Conflicts:
	topaz/objects/floatobject.py
	topaz/objspace.py

+++ lib-topaz/array.rb
@@ -11,11 +11,8 @@ class Array
         return self.replace(size_or_arr.to_ary)
       end
     end
-    if !size_or_arr.respond_to?(:to_int)
-      raise TypeError, "can't convert #{size_or_arr.class} into Integer"
-    end
-    length = size_or_arr.to_int
-    raise ArgumentError, "negative array size" if length < 0
+    length = Topaz.convert_type(size_or_arr, Fixnum, :to_int)
+    raise ArgumentError.new("negative array size") if length < 0
     if block
       # TODO: Emit "block supersedes default value argument" warning
       length.times { |i| self << yield(i) }
@@ -25,16 +22,27 @@ class Array
     return self
   end
 
-  def to_s
+  def self.[](*args)
+    args.inject(allocate) { |array, arg| array << arg}
+  end
+
+  def inspect
     result = "["
-    self.each_with_index do |obj, i|
-      if i > 0
-        result << ", "
+    recursion = Thread.current.recursion_guard(:array_inspect, self) do
+      self.each_with_index do |obj, i|
+        if i > 0
+          result << ", "
+        end
+        result << obj.inspect
       end
-      result << obj.to_s
+    end
+    if recursion
+      result << "..."
     end
     result << "]"
   end
+
+  alias :to_s :inspect
 
   def -(other)
     res = []
@@ -50,12 +58,25 @@ class Array
     self[idx]
   end
 
-  def each
+  def fetch(*args, &block)
+    i = Topaz.convert_type(args[0], Fixnum, :to_int)
+    if i < -self.length || i >= self.length
+      return block.call(args[0]) if block
+      return args[1] if args.size > 1
+      raise IndexError.new("index #{i} outside of array bounds: -#{self.length}...#{self.length}")
+    else
+      self[i]
+    end
+  end
+
+  def each(&block)
+    return self.enum_for(:each) unless block
     i = 0
     while i < self.length
       yield self[i]
       i += 1
     end
+    return self
   end
 
   def zip(ary)
@@ -85,6 +106,7 @@ class Array
   end
 
   def reject!(&block)
+    return self.enum_for(:reject!) unless block
     prev_size = self.size
     self.delete_if(&block)
     return nil if prev_size == self.size
@@ -99,8 +121,9 @@ class Array
     detect { |arr| arr.is_a?(Array) && arr[1] == value }
   end
 
-  def delete_if
-    raise RuntimeError, "can't modify frozen #{self.class}" if frozen?
+  def delete_if(&block)
+    raise RuntimeError.new("can't modify frozen #{self.class}") if frozen?
+    return self.enum_for(:delete_if) unless block
     i = 0
     c = 0
     sz = self.size
@@ -125,13 +148,17 @@ class Array
     return nil
   end
 
-  def first
-    return self[0]
+  def first(*args)
+    if args.empty?
+      self[0]
+    else
+      take(*args)
+    end
   end
 
   def flatten(level = -1)
     list = []
-    recursion = Thread.current.recursion_guard(self) do
+    Thread.current.recursion_guard(:array_flatten, self) do
       self.each do |item|
         if level == 0
           list << item
@@ -143,15 +170,12 @@ class Array
       end
       return list
     end
-    if recursion
-      raise ArgumentError, "tried to flatten recursive array"
-    end
+    raise ArgumentError.new("tried to flatten recursive array")
   end
 
   def flatten!(level = -1)
     list = self.flatten(level)
-    self.clear
-    return self.concat list
+    self.replace(list)
   end
 
   def sort(&block)
@@ -163,14 +187,20 @@ class Array
       return true
     end
     if !other.kind_of?(Array)
-      return false
+      if other.respond_to?(:to_ary)
+        return other == self
+      else
+        return false
+      end
     end
     if self.size != other.size
       return false
     end
-    self.each_with_index do |x, i|
-      if x != other[i]
-        return false
+    Thread.current.recursion_guard(:array_equals, self) do
+      self.each_with_index do |x, i|
+        if x != other[i]
+          return false
+        end
       end
     end
     return true
@@ -186,9 +216,11 @@ class Array
     if self.length != other.length
       return false
     end
-    self.each_with_index do |x, i|
-      if !x.eql?(other[i])
-        return false
+    Thread.current.recursion_guard(:array_eqlp, self) do
+      self.each_with_index do |x, i|
+        if !x.eql?(other[i])
+          return false
+        end
       end
     end
     return true
@@ -205,16 +237,8 @@ class Array
 
   def *(arg)
     return join(arg) if arg.respond_to? :to_str
-
-    # MRI error cases
-    argcls = arg.class
-    begin
-      arg = arg.to_int
-    rescue Exception
-      raise TypeError, "can't convert #{argcls} into Fixnum"
-    end
-    raise TypeError, "can't convert #{argcls} to Fixnum (argcls#to_int gives arg.class)" if arg.class != Fixnum
-    raise ArgumentError, "Count cannot be negative" if arg < 0
+    arg = Topaz.convert_type(arg, Fixnum, :to_int)
+    raise ArgumentError.new("Count cannot be negative") if arg < 0
 
     return [] if arg == 0
     result = self.dup
@@ -225,26 +249,21 @@ class Array
   end
 
   def map!(&block)
-    raise RuntimeError, "can't modify frozen #{self.class}" if frozen?
+    return self.enum_for(:map!) unless block
+    raise RuntimeError.new("can't modify frozen #{self.class}") if frozen?
     self.each_with_index { |obj, idx| self[idx] = yield(obj) }
     self
   end
 
-  def max(&block)
-    max = self[0]
-    self.each do |e|
-      max = e if (block ? block.call(max, e) : max <=> e) < 0
-    end
-    max
-  end
+  alias :collect! :map!
 
   def uniq!(&block)
-    raise RuntimeError, "can't modify frozen #{self.class}" if frozen?
+    raise RuntimeError.new("can't modify frozen #{self.class}") if frozen?
     seen = {}
-    old_len = self.length
+    old_length = self.length
     i = 0
     shifted = 0
-    while i < self.length do
+    while i < old_length do
       item = self[i]
       item = yield(item) if block
       if seen.include? item
@@ -271,5 +290,77 @@ class Array
 
   def values_at(*args)
     args.map { |n| self[n] }
+  end
+
+  def each_index(&block)
+    return self.enum_for(:each_index) unless block
+    0.upto(size - 1, &block)
+    self
+  end
+
+  def reverse
+    self.dup.reverse!
+  end
+
+  def reverse_each(&block)
+    return self.enum_for(:reverse_each) unless block
+    i = self.length - 1
+    while i >= 0
+      yield self[i]
+      i -= 1
+    end
+    return self
+  end
+
+  def index(obj = nil, &block)
+    return self.enum_for(:index) if !obj && !block
+    each_with_index do |e, i|
+      return i if obj ? (e == obj) : block.call(e)
+    end
+    nil
+  end
+
+  alias :find_index :index
+
+  def rindex(obj = nil, &block)
+    return self.enum_for(:rindex) if !obj && !block
+    reverse.each_with_index do |e, i|
+      return length - i - 1 if obj ? (e == obj) : block.call(e)
+    end
+    nil
+  end
+
+  def rotate(n = 1)
+    Array.new(self).rotate!(n)
+  end
+
+  def count(*args, &block)
+    c = 0
+    if args.empty?
+      if block
+        self.each { |e| c += 1 if block.call(e) }
+      else
+        c = self.length
+      end
+    else
+      arg = args[0]
+      self.each { |e| c += 1 if e == arg }
+    end
+    c
+  end
+
+  def shuffle!
+    raise RuntimeError.new("can't modify frozen #{self.class}") if frozen?
+    (self.length - 1).downto(1) do |idx|
+      other = rand(idx + 1)
+      self[other], self[idx] = self[idx], self[other]
+    end
+    self
+  end
+
+  def shuffle
+    arr = Array.new(self)
+    arr.shuffle!
+    arr
   end
 end
